# storage: implement intent resolution using LockTableIterator