@nvanbenschoten 2023-09-15
> Is the problem that the LockUpdate does not tell us what locks were held by this txn on the span?
> What if we tracked spans for each lock strength in the txn record?

Right, we don't currently track which locks with which strengths are held on which keys, only that there are locks with _some_ strength held on certain keys. Tracking the strength of locks in the coordinator and then storing locks with their strength in the txn record would let us perform more precise resolution.

> I can't remember whether this optimization was prompted by only microbenchmarks.

From [what I was able to dig up](https://github.com/cockroachdb/cockroach/pull/60622#issuecomment-780165629), it sounds like one of the SQL logic tests that performed ranged intent resolution became significantly slower with separated intents. I don't see any mention of single-key resolution being so slow that we had to optimize it. It makes me wonder whether the first bullet in the commit message was an opportunistic improvement.

I'm running the benchmarks introduced in that commit over this PR to get a better sense for the impact.

> Another possibility would be to do a series of multiple SeekGEWithLimit calls for each of the strengths. We don't fully optimize this monotonic seek case in pebble.Iterator but it is not hard to change that to use next calls under the hood. We would still need to produce a limit key, which is a bit problematic since it needs to be a valid EngineKey for cmp to work and needs to be exclusive. If we can efficiently construct the immediately smaller txnid, then we can produce a limit key.

I've been considering this. I would improve the worst-case performance of single-key resolution in exchange for a more expensive common case (3 seeks instead of 1). I wonder if that's worth it. We already scan over all locks on a key (including tombstones) when writing these intents, so the degenerate quadratic case where a single key is written and resolved in fast succession is quadratic no matter what we do here.