@nvanbenschoten 2023-09-23
> I'll capture the impact of this change on the following benchmarks and evaluate

Here is the impact that this change has on the benchmarks added in https://github.com/cockroachdb/cockroach/commit/d1c91e0e8180eec95c9365a101a1912649bb0dab. We see that it's non-impactful except for the most degenerate cases, as expected:
```
name                                                  old time/op  new time/op   delta
IntentScan/versions=10/percent-flushed=0-30           2.84µs ± 0%   2.82µs ± 0%    -0.78%  (p=0.000 n=10+9)
IntentScan/versions=10/percent-flushed=50-30          1.61µs ± 0%   1.60µs ± 0%    -1.19%  (p=0.000 n=10+10)
IntentScan/versions=10/percent-flushed=80-30          1.14µs ± 0%   1.13µs ± 0%    -1.21%  (p=0.000 n=10+9)
IntentScan/versions=10/percent-flushed=90-30           943ns ± 0%    934ns ± 0%    -0.95%  (p=0.000 n=10+10)
IntentScan/versions=10/percent-flushed=100-30          895ns ± 0%    888ns ± 0%    -0.75%  (p=0.000 n=10+10)
IntentScan/versions=100/percent-flushed=0-30          33.1µs ± 1%   31.8µs ± 2%    -4.15%  (p=0.000 n=8+10)
IntentScan/versions=100/percent-flushed=50-30         8.56µs ± 4%   8.37µs ± 3%      ~     (p=0.050 n=9+9)
IntentScan/versions=100/percent-flushed=80-30         4.29µs ± 2%   4.28µs ± 2%      ~     (p=0.796 n=10+10)
IntentScan/versions=100/percent-flushed=90-30         2.66µs ± 1%   2.65µs ± 2%      ~     (p=0.056 n=9+10)
IntentScan/versions=100/percent-flushed=100-30         975ns ± 1%    970ns ± 0%    -0.53%  (p=0.001 n=10+9)
IntentScan/versions=200/percent-flushed=0-30          67.5µs ± 1%   65.0µs ± 1%    -3.66%  (p=0.000 n=10+10)
IntentScan/versions=200/percent-flushed=50-30         20.3µs ± 1%   20.1µs ± 8%      ~     (p=0.400 n=9+10)
IntentScan/versions=200/percent-flushed=80-30         7.13µs ± 6%   6.87µs ± 2%    -3.65%  (p=0.003 n=10+10)
IntentScan/versions=200/percent-flushed=90-30         4.26µs ± 3%   4.24µs ± 2%      ~     (p=0.393 n=10+10)
IntentScan/versions=200/percent-flushed=100-30        1.04µs ± 1%   1.04µs ± 1%      ~     (p=0.135 n=10+10)
IntentScan/versions=400/percent-flushed=0-30          33.6µs ± 3%   33.0µs ± 2%    -1.89%  (p=0.001 n=10+9)
IntentScan/versions=400/percent-flushed=50-30         43.9µs ± 1%   43.2µs ± 1%    -1.66%  (p=0.002 n=10+10)
IntentScan/versions=400/percent-flushed=80-30         15.3µs ± 4%   15.4µs ± 7%      ~     (p=0.684 n=10+10)
IntentScan/versions=400/percent-flushed=90-30         7.14µs ± 2%   7.09µs ± 4%      ~     (p=0.161 n=9+9)
IntentScan/versions=400/percent-flushed=100-30        1.77µs ± 2%   1.75µs ± 1%    -0.88%  (p=0.023 n=9+9)
IntentResolution/versions=10/percent-flushed=0-30     3.60µs ± 1%   3.80µs ± 1%    +5.56%  (p=0.000 n=10+10)
IntentResolution/versions=10/percent-flushed=50-30    3.16µs ± 2%   2.55µs ± 1%   -19.27%  (p=0.000 n=10+10)
IntentResolution/versions=10/percent-flushed=80-30    2.61µs ± 1%   2.29µs ± 2%   -12.41%  (p=0.000 n=10+10)
IntentResolution/versions=10/percent-flushed=90-30    2.55µs ± 1%   2.07µs ± 1%   -18.69%  (p=0.000 n=10+9)
IntentResolution/versions=10/percent-flushed=100-30   2.44µs ± 1%   1.98µs ± 1%   -19.04%  (p=0.000 n=10+10)
IntentResolution/versions=100/percent-flushed=0-30    10.5µs ± 5%   18.2µs ± 3%   +73.37%  (p=0.000 n=10+10)
IntentResolution/versions=100/percent-flushed=50-30   4.54µs ± 1%   6.30µs ± 2%   +38.76%  (p=0.000 n=10+10)
IntentResolution/versions=100/percent-flushed=80-30   4.28µs ± 2%   3.84µs ± 1%   -10.26%  (p=0.000 n=10+10)
IntentResolution/versions=100/percent-flushed=90-30   3.90µs ± 1%   3.12µs ± 1%   -19.87%  (p=0.000 n=10+10)
IntentResolution/versions=100/percent-flushed=100-30  2.61µs ± 1%   2.03µs ± 1%   -22.08%  (p=0.000 n=10+9)
IntentResolution/versions=200/percent-flushed=0-30    14.8µs ± 4%   35.9µs ± 2%  +143.54%  (p=0.000 n=9+10)
IntentResolution/versions=200/percent-flushed=50-30   4.86µs ± 2%  10.53µs ± 5%  +116.63%  (p=0.000 n=10+10)
IntentResolution/versions=200/percent-flushed=80-30   4.60µs ± 2%   5.59µs ± 1%   +21.50%  (p=0.000 n=10+9)
IntentResolution/versions=200/percent-flushed=90-30   4.35µs ± 2%   3.90µs ± 1%   -10.32%  (p=0.000 n=10+10)
IntentResolution/versions=200/percent-flushed=100-30  2.65µs ± 1%   2.07µs ± 1%   -21.89%  (p=0.000 n=10+10)
IntentResolution/versions=400/percent-flushed=0-30    5.10µs ± 2%  16.54µs ± 3%  +224.50%  (p=0.000 n=10+10)
IntentResolution/versions=400/percent-flushed=50-30   5.16µs ± 2%  22.19µs ± 2%  +329.77%  (p=0.000 n=10+10)
IntentResolution/versions=400/percent-flushed=80-30   4.92µs ± 0%   8.74µs ± 3%   +77.82%  (p=0.000 n=8+10)
IntentResolution/versions=400/percent-flushed=90-30   4.67µs ± 1%   5.60µs ± 1%   +19.80%  (p=0.000 n=10+10)
IntentResolution/versions=400/percent-flushed=100-30  2.83µs ± 1%   2.10µs ± 1%   -25.91%  (p=0.000 n=10+10)
```
Given that these cases impact the front half of mutations regardless of what we do during intent resolution, I'm ok with the limited regressions on this microbenchmark. Still, to minimize the impact, I re-introduced half of the previous optimization into `MVCCResolveWriteIntent` by initially seeking to the intent version instead of the version prefix. I also left a comment explaining the opportunity to optimize further if we track the strengths that locks on a given key are held with, with a reference back to d1c91e0e. For now, I'll avoid optimizing further until we see this come up in real-world workloads.

TFTRs!

bors r+