@nvanbenschoten 2023-09-25
For completeness, here's the impact of this change on `BenchmarkIntentRangeResolution`:
```
name                                                                                               old time/op  new time/op  delta
IntentRangeResolution/versions=10/sparseness=1/other-txn-intents=false/percent-flushed=0-30         256µs ± 1%   263µs ± 2%   +2.95%  (p=0.000 n=10+10)
IntentRangeResolution/versions=10/sparseness=1/other-txn-intents=false/percent-flushed=50-30        117µs ± 1%   124µs ± 1%   +6.11%  (p=0.000 n=10+9)
IntentRangeResolution/versions=10/sparseness=1/other-txn-intents=false/percent-flushed=100-30      68.0µs ± 2%  75.4µs ± 2%  +10.89%  (p=0.000 n=10+10)
IntentRangeResolution/versions=10/sparseness=100/other-txn-intents=false/percent-flushed=0-30       210µs ± 1%   208µs ± 1%   -0.78%  (p=0.000 n=10+9)
IntentRangeResolution/versions=10/sparseness=100/other-txn-intents=false/percent-flushed=50-30     51.6µs ± 1%  51.5µs ± 1%     ~     (p=0.905 n=9+10)
IntentRangeResolution/versions=10/sparseness=100/other-txn-intents=false/percent-flushed=100-30    2.19µs ± 4%  2.28µs ± 2%   +4.09%  (p=0.000 n=10+9)
IntentRangeResolution/versions=10/sparseness=100/other-txn-intents=true/percent-flushed=0-30        237µs ± 1%   203µs ± 0%  -14.13%  (p=0.000 n=9+10)
IntentRangeResolution/versions=10/sparseness=100/other-txn-intents=true/percent-flushed=50-30      81.9µs ± 1%  50.7µs ± 1%  -38.11%  (p=0.000 n=9+9)
IntentRangeResolution/versions=10/sparseness=100/other-txn-intents=true/percent-flushed=100-30     42.1µs ± 1%  12.4µs ± 1%  -70.53%  (p=0.000 n=9+10)
IntentRangeResolution/versions=10/sparseness=1000/other-txn-intents=false/percent-flushed=0-30      209µs ± 1%   206µs ± 1%   -1.72%  (p=0.000 n=10+9)
IntentRangeResolution/versions=10/sparseness=1000/other-txn-intents=false/percent-flushed=50-30    49.7µs ± 1%  49.8µs ± 1%     ~     (p=0.447 n=10+9)
IntentRangeResolution/versions=10/sparseness=1000/other-txn-intents=false/percent-flushed=100-30   1.34µs ± 3%  1.32µs ± 4%     ~     (p=0.051 n=9+10)
IntentRangeResolution/versions=10/sparseness=1000/other-txn-intents=true/percent-flushed=0-30       236µs ± 1%   202µs ± 1%  -14.17%  (p=0.000 n=10+10)
IntentRangeResolution/versions=10/sparseness=1000/other-txn-intents=true/percent-flushed=50-30     81.8µs ± 1%  49.6µs ± 1%  -39.32%  (p=0.000 n=10+10)
IntentRangeResolution/versions=10/sparseness=1000/other-txn-intents=true/percent-flushed=100-30    41.8µs ± 1%  11.5µs ± 5%  -72.35%  (p=0.000 n=9+10)
IntentRangeResolution/versions=100/sparseness=1/other-txn-intents=false/percent-flushed=0-30       2.77ms ± 1%  2.77ms ± 2%     ~     (p=0.604 n=9+10)
IntentRangeResolution/versions=100/sparseness=1/other-txn-intents=false/percent-flushed=50-30       717µs ± 4%   735µs ± 4%   +2.57%  (p=0.035 n=10+10)
IntentRangeResolution/versions=100/sparseness=1/other-txn-intents=false/percent-flushed=100-30     67.3µs ± 1%  74.6µs ± 1%  +10.85%  (p=0.000 n=9+8)
IntentRangeResolution/versions=100/sparseness=100/other-txn-intents=false/percent-flushed=0-30     2.74ms ± 1%  2.75ms ± 2%     ~     (p=0.481 n=10+10)
IntentRangeResolution/versions=100/sparseness=100/other-txn-intents=false/percent-flushed=50-30     651µs ± 6%   632µs ± 3%   -2.99%  (p=0.028 n=10+9)
IntentRangeResolution/versions=100/sparseness=100/other-txn-intents=false/percent-flushed=100-30   2.13µs ± 0%  2.28µs ± 2%   +7.08%  (p=0.000 n=8+10)
IntentRangeResolution/versions=100/sparseness=100/other-txn-intents=true/percent-flushed=0-30      2.74ms ± 1%  2.73ms ± 2%     ~     (p=0.247 n=10+10)
IntentRangeResolution/versions=100/sparseness=100/other-txn-intents=true/percent-flushed=50-30      674µs ± 2%   622µs ± 2%   -7.64%  (p=0.000 n=10+9)
IntentRangeResolution/versions=100/sparseness=100/other-txn-intents=true/percent-flushed=100-30    39.5µs ± 0%  12.4µs ± 1%  -68.70%  (p=0.000 n=8+9)
IntentRangeResolution/versions=100/sparseness=1000/other-txn-intents=false/percent-flushed=0-30    2.73ms ± 2%  2.72ms ± 2%     ~     (p=0.739 n=10+10)
IntentRangeResolution/versions=100/sparseness=1000/other-txn-intents=false/percent-flushed=50-30    639µs ± 6%   628µs ± 6%     ~     (p=0.280 n=10+10)
IntentRangeResolution/versions=100/sparseness=1000/other-txn-intents=false/percent-flushed=100-30  1.35µs ± 3%  1.32µs ± 2%   -2.82%  (p=0.000 n=10+9)
IntentRangeResolution/versions=100/sparseness=1000/other-txn-intents=true/percent-flushed=0-30     2.76ms ± 2%  2.70ms ± 1%   -2.15%  (p=0.000 n=10+9)
IntentRangeResolution/versions=100/sparseness=1000/other-txn-intents=true/percent-flushed=50-30     670µs ± 3%   620µs ± 4%   -7.44%  (p=0.000 n=9+10)
IntentRangeResolution/versions=100/sparseness=1000/other-txn-intents=true/percent-flushed=100-30   39.1µs ± 1%  11.5µs ± 1%  -70.63%  (p=0.000 n=8+10)
IntentRangeResolution/versions=400/sparseness=1/other-txn-intents=false/percent-flushed=0-30       3.06ms ± 1%  3.07ms ± 2%     ~     (p=0.315 n=10+10)
IntentRangeResolution/versions=400/sparseness=1/other-txn-intents=false/percent-flushed=50-30      4.08ms ± 1%  4.08ms ± 1%     ~     (p=0.842 n=9+10)
IntentRangeResolution/versions=400/sparseness=1/other-txn-intents=false/percent-flushed=100-30     70.8µs ± 2%  78.3µs ± 2%  +10.63%  (p=0.000 n=10+10)
IntentRangeResolution/versions=400/sparseness=100/other-txn-intents=false/percent-flushed=0-30     2.99ms ± 2%  2.99ms ± 3%     ~     (p=0.739 n=10+10)
IntentRangeResolution/versions=400/sparseness=100/other-txn-intents=false/percent-flushed=50-30    3.99ms ± 2%  4.00ms ± 2%     ~     (p=0.436 n=10+10)
IntentRangeResolution/versions=400/sparseness=100/other-txn-intents=false/percent-flushed=100-30   2.20µs ± 3%  2.32µs ± 2%   +5.55%  (p=0.000 n=9+10)
IntentRangeResolution/versions=400/sparseness=100/other-txn-intents=true/percent-flushed=0-30      2.99ms ± 1%  2.95ms ± 2%   -1.57%  (p=0.000 n=10+10)
IntentRangeResolution/versions=400/sparseness=100/other-txn-intents=true/percent-flushed=50-30     4.03ms ± 2%  3.96ms ± 2%   -1.81%  (p=0.004 n=10+9)
IntentRangeResolution/versions=400/sparseness=100/other-txn-intents=true/percent-flushed=100-30    43.2µs ± 2%  12.1µs ± 1%  -71.91%  (p=0.000 n=9+10)
IntentRangeResolution/versions=400/sparseness=1000/other-txn-intents=false/percent-flushed=0-30    2.98ms ± 2%  2.95ms ± 1%     ~     (p=0.052 n=10+10)
IntentRangeResolution/versions=400/sparseness=1000/other-txn-intents=false/percent-flushed=50-30   3.94ms ± 0%  3.93ms ± 1%     ~     (p=0.481 n=8+9)
IntentRangeResolution/versions=400/sparseness=1000/other-txn-intents=false/percent-flushed=100-30  1.37µs ± 2%  1.33µs ± 0%   -2.95%  (p=0.000 n=9+8)
IntentRangeResolution/versions=400/sparseness=1000/other-txn-intents=true/percent-flushed=0-30     3.00ms ± 1%  2.95ms ± 3%   -1.92%  (p=0.001 n=10+10)
IntentRangeResolution/versions=400/sparseness=1000/other-txn-intents=true/percent-flushed=50-30    3.99ms ± 1%  3.94ms ± 2%   -1.24%  (p=0.006 n=9+10)
IntentRangeResolution/versions=400/sparseness=1000/other-txn-intents=true/percent-flushed=100-30   42.8µs ± 3%  11.2µs ± 2%  -73.74%  (p=0.000 n=9+9)
```